Improve `pytest` flow
