--- community/modules/scheduler/schedmd-slurm-gcp-v6-controller/modules/slurm_files/scripts/tests/README.md
@@ -1,6 +1,0 @@
-# Unit tests
-
-```sh
-# cwd is scripts/tests
-$ pytest -W ignore::DeprecationWarning
-```
