+++ community/modules/scheduler/schedmd-slurm-gcp-v6-controller/modules/slurm_files/scripts/tests/test_topology.py
@@ -12,50 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Optional
 import mock
-import sys
+from common import TstCfg, TstNodeset, TstTPU, make_to_hostnames_mock
 
-if ".." not in sys.path:
-    sys.path.append("..")  # TODO: make this more robust
 import util
 import conf
-
-from dataclasses import dataclass, field
 import tempfile
-
-
-# TODO: use "real" classes once they are defined (instead of NSDict)
-@dataclass
-class TstNodeset:
-    nodeset_name: str
-    node_count_static: int = 0
-    node_count_dynamic_max: int = 0
-
-
-@dataclass
-class TstCfg:
-    slurm_cluster_name: str = "m22"
-    nodeset: dict[str, TstNodeset] = field(default_factory=dict)
-    nodeset_tpu: dict[str, TstNodeset] = field(default_factory=dict)
-    output_dir: Optional[str] = None
-
-
-@dataclass
-class TstTPU:  # to prevent client initialization durint "TPU.__init__"
-    vmcount: int
-
-
-def make_to_hostnames_mock(tbl: Optional[dict[str, list[str]]]):
-    tbl = tbl or {}
-
-    def se(k: str) -> list[str]:
-        if k not in tbl:
-            raise AssertionError(f"to_hostnames mock: unexpected nodelist: '{k}'")
-        return tbl[k]
-
-    return se
-
 
 def test_gen_topology_conf_empty():
     cfg = TstCfg(output_dir=tempfile.mkdtemp())
